add task to clean UserFile
