+++ django_project/gap_api/tasks/__init__.py
@@ -1,2 +1,2 @@
 from gap_api.tasks.api_log import *  # noqa
-from gap_api.tasks.cleanup_location import *  # noqa
+from django_project.gap_api.tasks.cleanup import *  # noqa
